Fix ISP jumper stickiness, nbytes wrap, and 256-byte page counter.
JP3 now pins 8 kHz for the whole CONNECT session; pagecounter is 16-bit so mega128 pages flush; usbisp boards document that JP3 is ignored.

+++ firmware/include/usbasp/prog_state.h
@@ -11,7 +11,7 @@ extern unsigned long prog_address;
 extern unsigned int prog_nbytes;
 extern unsigned int prog_pagesize;
 extern uchar prog_blockflags;
-extern uchar prog_pagecounter;
+extern unsigned int prog_pagecounter;
 extern uchar replyBuffer[8];
 
 usbMsgLen_t usbasp_vendor_setup(uchar data[8]);
